update rectangle exercise

+++ src/Rectangle.java
@@ -6,12 +6,12 @@ public class Rectangle {
 
         Scanner input = new Scanner(System.in);
         System.out.print("Nhập chiều cao: ");
-        int height = input.nextInt();
+        float height = input.nextFloat();
         System.out.print("Nhập chiều rộng : ");
-        int width = input.nextInt();
+        float width = input.nextFloat();
 
-        int perimeter = (height + width) * 2;
-        int area = height * width;
+        float perimeter = (height + width) * 2;
+        float area = height * width;
 
         System.out.println("Chu vi hình chữ nhật : " + perimeter);
         System.out.println("Diện tích hình chữ nhật : " + area);
